fix TMIDT parsing
