+++ configs/preprocess/TMIDT.yaml
@@ -5,10 +5,10 @@ shared:
   sample_rate: 44100
 
 EvalDataPreprocess:
-  dataset_path: "/Users/pierfrancescomelucci/Research/Playground/Sony/midi"
-  output_path: "/Users/pierfrancescomelucci/Research/Playground/Sony/data"
+  dataset_path: "path_to_dataset"
+  output_path: "path_to_output"
   dataset_name: "TMIDT"
   audio_extension: "mp3"
-  dataset_size: "m"
+  dataset_size: "medium"
   drums_only: true
-  num_audio_files: 100
+  num_audio_files: null # debug only
