+++ data_modules/text_parser.py
@@ -164,6 +164,8 @@ class TMIDTTextParser(DrumTextParser):
                 print(e)
                 continue
             for audio_chunk, notes_chunk in zip(audio_chunks, notes_chunks):
+                if not len(notes_chunk):
+                    continue
                 wav_bin = audio_chunk.numpy().astype(np.float32).tobytes()
                 notes_bin = np.array(notes_chunk, dtype=np.float32).tobytes()
                 batch_rows["audio_id"].append(audio_id)
