+++ data_modules/train_dataset.py
@@ -21,19 +21,21 @@ from modules.midi_tokenizer import MidiTokenizer, MidiTokenizerConfig
 
 
 @dataclass
-class LakhDatasetConfig(SharedConfig):
+class TrainDatasetConfig(SharedConfig):
     dataset_path: str
     empty_tokens_percentage: float
-    partitions: list[str]
     random_velocity_prob: float
+    dataset_name: str
 
 
 @dataclass
-class TMIDTDatasetConfig(SharedConfig):
-    dataset_path: str
-    empty_tokens_percentage: float
+class LakhDatasetConfig(TrainDatasetConfig):
     partitions: list[str]
-    random_velocity_prob: float
+
+
+@dataclass
+class TMIDTDatasetConfig(TrainDatasetConfig):
+    pass
 
 
 def collate_fn(batch):
@@ -54,25 +56,15 @@ def collate_fn(batch):
     }
 
 
-class TMIDTDataset(Dataset):
-    def __init__(self, config: TMIDTDatasetConfig, tokenizer: MidiTokenizer):
+class TrainDataset(Dataset):
+    def __init__(self, config: TrainDatasetConfig, tokenizer: MidiTokenizer):
         self.config = config
-        self.sample_rate = config.sample_rate
-        self.dataset_path = config.dataset_path
-        self.dataset = load_dataset(
-            "parquet",
-            data_files=config.dataset_path,
-        )
         self.tokenizer = tokenizer
+        self.empty_tokens_percentage = config.empty_tokens_percentage
+        self.sample_rate = config.sample_rate
 
-    def __len__(self):
-        return len(self.dataset)
-
-    def _empty_wav(self):
-        return torch.zeros(int(self.config.input_sec * self.config.sample_rate))
-
-    def _binary_to_torch(self, binary_data):
-        return torch.from_numpy(np.frombuffer(binary_data, dtype=np.float32))
+    def _resample_audio(self, audio: torch.Tensor, sr: int, target_sr: int):
+        return torchaudio.transforms.Resample(sr, target_sr)(audio)
 
     def _map_notes_to_Gm_custom(self, example):
         notes = torch.stack(self._binary_to_torch(example["notes"]).split(4))  # onset, offset, pitch, velocity
@@ -81,6 +73,68 @@ class TMIDTDataset(Dataset):
 
     def _tokenize_notes(self, example):
         example["tokens"] = self.tokenizer.notes_to_adt_tokens(example["notes"])
+        return example
+
+    def __len__(self):
+        return len(self.dataset)
+
+    def _binary_to_torch(self, binary_data):
+        return torch.from_numpy(np.frombuffer(binary_data, dtype=np.float32))
+
+    def _empty_wav(self):
+        return torch.zeros(int(self.config.input_sec * self.config.sample_rate))
+
+    def __getitem__(self, index):
+        pass
+
+    def get_dataloader(self, batch_size: int, shuffle: bool, num_workers: int):
+        return DataLoader(
+            self,
+            batch_size=batch_size,
+            shuffle=shuffle,
+            collate_fn=collate_fn,
+            num_workers=num_workers,
+        )
+
+
+class TMIDTDataset(TrainDataset):
+    def __init__(self, config: TMIDTDatasetConfig, tokenizer: MidiTokenizer):
+        super().__init__(config, tokenizer)
+        self.config = config
+        self.sample_rate = config.sample_rate
+        self.dataset_path = config.dataset_path
+        self.dataset = load_dataset(
+            config.dataset_path,
+        )["train"]
+        self.tokenizer = tokenizer
+        num_proc = 1  # min(32, os.cpu_count())
+
+        if not config.random_velocity_prob:
+            # map notes to Gm_custom
+            self.dataset = self.dataset.map(self._map_notes_to_Gm_custom, num_proc=num_proc)
+            # convert notes to tokens
+            self.dataset = self.dataset.map(self._tokenize_notes, num_proc=num_proc)
+        else:
+            logging.info(
+                "When the flag random_velocity_prob is True, the velocity will be randomly generated each time when an item is drawn, and the tokenization preprocess will be skipped in favour of a on-the-fly tokenization."
+            )
+
+        self.dataset = self.dataset.map(self._get_audio, num_proc=num_proc)
+
+    def _map_notes_to_Gm_custom(
+        self, example
+    ):  # we need to define here this method otherwise the map method will not work
+        notes = torch.from_numpy(np.frombuffer(example["notes"], dtype=np.float32))
+        notes = torch.stack(notes.split(4))  # onset, offset, pitch, velocity
+        example["notes"] = self.tokenizer.map_notes_to_Gm_custom(notes)
+        return example
+
+    def _get_audio(self, example):
+        data = torch.from_numpy(np.frombuffer(example["audio"], dtype=np.float32))
+        resampler = torchaudio.transforms.Resample(44100, self.sample_rate)
+        audio = resampler(data)
+        # normalize audio
+        example["audio"] = audio / (audio.abs().max() + 1e-8)
         return example
 
     def __getitem__(self, index):
@@ -98,7 +152,7 @@ class TMIDTDataset(Dataset):
         else:
             notes = item["notes"]
             tokens = item["tokens"]
-        wavs = item["audio"]
+        wavs = torch.tensor(item["audio"])
         return wavs, tokens
 
     def get_dataloader(self, batch_size: int, shuffle: bool, num_workers: int):
@@ -146,24 +200,6 @@ class LakhDataset(Dataset):
             )
         self.empty_tokens_percentage = config.empty_tokens_percentage
 
-    def _map_notes_to_Gm_custom(self, example):
-        notes = torch.stack(self._binary_to_torch(example["notes"]).split(4))  # onset, offset, pitch, velocity
-        example["notes"] = self.tokenizer.map_notes_to_Gm_custom(notes)
-        return example
-
-    def _tokenize_notes(self, example):
-        example["tokens"] = self.tokenizer.notes_to_adt_tokens(example["notes"])
-        return example
-
-    def __len__(self):
-        return len(self.dataset)
-
-    def _binary_to_torch(self, binary_data):
-        return torch.from_numpy(np.frombuffer(binary_data, dtype=np.float32))
-
-    def _empty_wav(self):
-        return torch.zeros(int(self.config.input_sec * self.config.sample_rate))
-
     def __getitem__(self, index):
         if random.random() < self.empty_tokens_percentage:
             return self._empty_wav(), self.tokenizer.empty_adt_tokens()
@@ -182,15 +218,6 @@ class LakhDataset(Dataset):
         wavs = self.synthetiser(notes)
         return wavs, tokens
 
-    def get_dataloader(self, batch_size: int, shuffle: bool, num_workers: int):
-        return DataLoader(
-            self,
-            batch_size=batch_size,
-            shuffle=shuffle,
-            collate_fn=collate_fn,
-            num_workers=num_workers,
-        )
-
 
 parser = argparse.ArgumentParser()
 parser.add_argument("config", type=str)
@@ -199,19 +226,29 @@ if __name__ == "__main__":
     # config
     config = load_config_from_yaml(args.config)
     config_tokenizer = config["tokenizer"]
-    config_synthetiser = config["synthetiser"]
-    config_synthetiser["ADTOF_mapping"] = config_tokenizer["ADTOF_mapping"]
-    config_synthetiser.update(config["shared"])
-    config_dataset = config["LakhDatasetConfig"]
+    config_dataset = config["TrainDatasetConfig"]
     config_dataset.update(config["shared"])
+
+    if config_dataset["dataset_name"] == "Lakh":
+        config_synthetiser = config.get("synthetiser", None)
+        assert config_synthetiser is not None, "Synthetiser is required for Lakh dataset"
+        config_synthetiser["ADTOF_mapping"] = config_tokenizer["ADTOF_mapping"]
+        config_synthetiser.update(config["shared"])
+        synthetiser = SynthDrum(SynthDrumConfig(**config_synthetiser))
+    else:
+        config_synthetiser = None
     # load modules
-    synthetiser = SynthDrum(SynthDrumConfig(**config_synthetiser))
     tokenizer = MidiTokenizer(MidiTokenizerConfig(**config_tokenizer))
-    dataset = LakhDataset(LakhDatasetConfig(**config_dataset), tokenizer, synthetiser)
+    if config_dataset["dataset_name"] == "Lakh":
+        dataset = LakhDataset(LakhDatasetConfig(**config_dataset), tokenizer, synthetiser)
+    elif config_dataset["dataset_name"] == "TMIDT":
+        dataset = TMIDTDataset(TMIDTDatasetConfig(**config_dataset), tokenizer)
+    else:
+        raise ValueError(f"Dataset name {config_dataset['dataset_name']} not supported")
     dataloader = dataset.get_dataloader(batch_size=4, shuffle=False, num_workers=4)
 
     counter = 0
-    sanity_check_path = "path_to_sanity_check"
+    sanity_check_path = "/Users/pierfrancescomelucci/Research/Playground/Sony/ADT_STR/sanity_check"
     os.makedirs(sanity_check_path, exist_ok=True)
     for batch in tqdm(dataloader, desc="Processing dataset"):
         # save wav in sanity check using torchaudio
